requirements.txt
streamlit
requests
azure.cognitiveservices.speech